+++ backend/src/app/models/Leitor.js
@@ -53,10 +53,6 @@ class Leitor {
       .first()
       .then((leitor) => leitor);
   }
-
-  async listaFaculdades(){
-    return await conexaoDb("faculdades").first().then((faculdades)=>faculdades)
-  }
 }
 
 export default Leitor;
